requirements.txt
Django==3.0.4
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
psycopg2-binary==2.8.3
gunicorn==20.0.4
pyjwt==1.7